Improve messages Media Scanner/ Hotplug
- move print to avoid a bad session debug:
[MediaScanner] scanning ('Card  Reader', '/media/usb', <__main__.Session
object at 0xafdea9d0>)
-Hotplug: add exception class CannotListenError(BindError)

thanks @norhap

+++ lib/python/Plugins/SystemPlugins/Hotplug/plugin.py
@@ -91,12 +91,13 @@ def autostart(reason, **kwargs):
 	if reason == 0:
 		from twisted.internet import reactor
 		try:
-			os.remove("/tmp/hotplug.socket")
-		except OSError:
-			pass
-		factory = Factory()
-		factory.protocol = Hotplug
-		reactor.listenUNIX("/tmp/hotplug.socket", factory)
+			if os.path.exists("/tmp/hotplug.socket"):
+				os.remove("/tmp/hotplug.socket")
+			factory = Factory()
+			factory.protocol = Hotplug
+			reactor.listenUNIX("/tmp/hotplug.socket", factory)
+		except (OSError, CannotListenError) as err:
+			print("[Hotplug]", err)
 
 
 def Plugins(**kwargs):
